Introduced necessities for puncher control
Slightly edited main.cpp to improve formatting. Added puncher.hpp and puncher.cpp which will be used to store the puncher based functions. Added comments and function headers to prepare for future implementation of puncher controls. Added a struct for controlling puncher via PID values and established puncher with that struct in globals.